plugin.yml: set api-version to 1.13
This provides free compatibility with 1.13+ servers as previously it has stopped the server from running BlackWidow with a version older than 1.21 even if the code itself works fine.

This removes the protection for older servers but server owners are already taking a risk there as the Wiki and plugin pages clearly advertise 1.21+ supported only and older versions are work on a 'good luck' basis.